requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
